Isolate container volume
Before this commit, the container's app directory was a binding to the
local project directory. This led to some confusion related to caching,
package installations and some directory creations as well:
- Newly installed package weren't seen inside the container until a lot
  of cache removals (which shouldn't happen in an isolated environment -
  hence revealing it wasn't isolated at all).
- When the container was started, `node_modules` and gatsby's `.cache`
  and `public` directories were generated with `root` as owner/group,
  which is very likely to be unwanted, as it would mean that, to simply
  delete the project directory, root permission would be needed.

This commit fixes both issues by:
- Letting the container's app directory in an isolated volume of its
  own;
- Copying package{,-lock}.json files to container's app directory;
- Creating a readonly binding for `src` directory.
This way, container's changes will happen only for the container's point
of view, with no visible local changes and without interferring in the
project's directory or other image builds (say, between branch changes).